requirements.txt
absl-py==2.3.1
appnope==0.1.4
asttokens==3.0.0
bleach==6.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.4
colorlog==4.7.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
duckdb==1.3.2
executing==2.2.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
grpcio==1.73.1
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kagglehub==0.3.12
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.48.0
nest-asyncio==1.6.0
networkx==3.5
numpy<2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
powerlaw==1.5
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
recbole==1.2.0
requests==2.32.4
scikit-learn==1.7.1
scikit-surprise==1.1.4
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
